% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfExtR_getDataMatrix.R
\name{zipfExtR_getDataMatrix}
\alias{zipfExtR_getDataMatrix}
\title{Convert a sample vector to a frequency matrix.}
\usage{
zipfExtR_getDataMatrix(values)
}
\arguments{
\item{values}{Vector of positive integer values.}
}
\value{
The matrix of frequencies associated to the vector \code{values}.
}
\description{
Converts a sequence of values into a matrix of frequencies.
}
\examples{
data <- rmoezipf(100, 2.5, 1.3)
zipfExtR_getDataMatrix(data)

}
