% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfpeFit.R
\name{zipfpeFit}
\alias{zipfpeFit}
\alias{residuals.zpeR}
\alias{fitted.zpeR}
\alias{coef.zpeR}
\alias{plot.zpeR}
\alias{print.zpeR}
\alias{summary.zpeR}
\alias{logLik.zpeR}
\alias{AIC.zpeR}
\alias{BIC.zpeR}
\title{Zipf-PE parameters estimation.}
\usage{
zipfpeFit(data, init_alpha, init_beta, level = 0.95, ...)

\method{residuals}{zpeR}(object, ...)

\method{fitted}{zpeR}(object, ...)

\method{coef}{zpeR}(object, ...)

\method{plot}{zpeR}(x, ...)

\method{print}{zpeR}(x, ...)

\method{summary}{zpeR}(object, ...)

\method{logLik}{zpeR}(object, ...)

\method{AIC}{zpeR}(object, ...)

\method{BIC}{zpeR}(object, ...)
}
\arguments{
\item{data}{Matrix of count data in form of table of frequencies.}

\item{init_alpha}{Initial value of \eqn{\alpha} parameter (\eqn{\alpha > 1}).}

\item{init_beta}{Initial value of \eqn{\beta} parameter (\eqn{\beta \in (-\infty, +\infty)}).}

\item{level}{Confidence level used to calculate the confidence intervals (default 0.95).}

\item{...}{Further arguments to the generic functions.The extra arguments are passing
to the \emph{\link{optim}} function.}

\item{object}{An object from class "zpeR" (output of \emph{zipfpeFit} function).}

\item{x}{An object from class "zpeR" (output of \emph{zipfpeFit} function).}
}
\value{
Returns an object composed by the maximum likelihood parameter estimations, their standard deviation, their confidence
intervals and the value of the log-likelihood at the maximum likelihood estimations.
}
\description{
For a given sample of strictly positive integer values,  usually of the type of ranking data or
frequencies of frequencies data, estimates the parameters of a Zipf-PE
distribution by means of the maximum likelihood method.
}
\details{
The argument \code{data} is a two column matrix such that the first column of each row contains a
count, while the corresponding second column contains its frequency.

The log-likelihood function is equal to:

\deqn{l(\alpha, \beta; x) = \beta\, (N - \zeta(\alpha)^{-1}\, \sum_{i = 1} ^m  f_{a}(x_{i}) \zeta(\alpha, x_i)) +
\sum_{i = 1} ^m f_{a}(x_{i})  log(e^{\frac{\beta\, x_{i}^{-\alpha}}{\zeta(\alpha)}} - 1) - N\, log(e^{\beta} - 1), }
where \eqn{m} is the number of different values in the sample, \eqn{N} is the sample size,
i.e.  \eqn{N = \sum_{i = 1} ^m x_i f_a(x_i)},  being \eqn{f_{a}(x_i)} is the absolute
frequency of \eqn{x_i}.

The function \emph{\link{optim}} is used to estimate the parameters.
}
\examples{
data <- rzipfpe(100, 2.5, 1.3)
data <- as.data.frame(table(data))
obj <- zipfpeFit(data, 1.1, 0.1)
}
\seealso{
\code{\link{moezipf_getInitialValues}}.
}
