% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfpssVariance.R
\name{zipfpssVariance}
\alias{zipfpssVariance}
\title{Variance of the Zipf-PSS distribution.}
\usage{
zipfpssVariance(alpha, lambda, isTruncated = FALSE)
}
\arguments{
\item{alpha}{Value of the \eqn{\alpha} parameter (\eqn{\alpha > 3}).}

\item{lambda}{Value of the \eqn{\lambda} parameter (\eqn{\lambda \geq 0}).}

\item{isTruncated}{Logical; if TRUE Use the zero-truncated version of the distribution to calculate the expected value (default = FALSE).}
}
\value{
A positive real value corresponding to the variance of the distribution.
}
\description{
Computes the variance of the Zipf-PSS distribution for given values of parameters
\eqn{\alpha} and \eqn{\lambda}.
}
\details{
The variance of the Zipf-PSS distribution only exists for \eqn{\alpha} values strictly greater than 3.
The value is derived from \eqn{Var[Y] = E[N]\, Var[X] + E[X]^2 \, Var[N]} where E[X] and
E[N] are the expected value of the Zipf and the Poisson distributions respectively.
In the same way the values of Var[X] and Var[N] stand for the variances of the Zipf and
the Poisson distributions. The resulting expression is set to be equal to:

\deqn{Var[Y] = \lambda\, \frac{\zeta(\alpha - 2)}{\zeta(\alpha)}}
Particularlly, the variance of the zero-truncated version of the Zipf-PSS distribution is
calculated as:
\deqn{Var[Y^{ZT}] = \frac{\lambda\, \zeta(\alpha)\, \zeta(\alpha - 2)\, (1 - e^{-\lambda}) - \lambda^2 \, \zeta(\alpha - 1)^2 \, e^{-\lambda}}{\zeta(\alpha)^2 \, (1 - e^{-\lambda})^2}}
}
\examples{
zipfpssVariance(4.5, 2.3)
zipfpssVariance(4.5, 2.3, TRUE)
}
\references{
{
Sarabia Alegría, JM. and Gómez Déniz, E. and Vázquez Polo, F. Estadística actuarial: teoría y aplicaciones. Pearson Prentice Hall.
}
}
