% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpe.R
\name{zpe}
\alias{zpe}
\alias{dzpe}
\alias{dzpe}
\alias{pzpe}
\title{The Zipf-Poisson Extreme Distribution (ZPE).}
\usage{
dzpe(x, alpha, lambda)

pzpe(q, alpha, beta)
}
\description{
Probability Mass Function, Cumulative Function of the ZPE distribution
with parameter \eqn{\alpha} and \eqn{\beta}.
}
