% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpeMean.R
\name{zpeMean}
\alias{zpeMean}
\title{Expected value of the ZPE distribution.}
\usage{
zpeMean(alpha, beta, tolerance = 10^(-4))
}
\arguments{
\item{alpha}{Value of the \eqn{\alpha} parameter (\eqn{\alpha > 2}).}

\item{beta}{Value of the \eqn{\beta} parameter (\eqn{\beta \in [-\infty, +\infty]}).}

\item{tolerance}{Tolerance used in the calculations (default = \eqn{10^{-4}}).}
}
\value{
A positive real value corresponding to the mean value of the ZPE distribution.
}
\description{
Computes the expected value of the ZPE distribution for given values of parameters
\eqn{\alpha} and \eqn{\beta}.
}
\details{
The mean of the distribution only exists for \eqn{\alpha} strictly greater than 2. It is computed
calculating the partial sums of the serie, and it stops when two consecutive partial
sums differs less than the \code{tolerance} value. The last partial sum is returned.
}
\examples{
zpeMean(2.5, 1.3)
zpeMean(2.5, 1.3, 10^(-3))
}
