% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpssMean.R
\name{zpssMean}
\alias{zpssMean}
\title{Expected value of the Z-PSS distribution.}
\usage{
zpssMean(alpha, lambda, isTruncated = FALSE)
}
\arguments{
\item{alpha}{Value of the \eqn{\alpha} parameter (\eqn{\alpha > 2}).}

\item{lambda}{Value of the \eqn{\lambda} parameter (\eqn{\lambda > 0}).}

\item{isTruncated}{Logical; if TRUE Use the zero-truncated version of the distribution to calculate the expected value (default = FALSE).}
}
\value{
A positive real value corresponding to the mean value of the distribution.
}
\description{
Computes the expected value of the Z-PSS distribution for given values of parameters.
\eqn{\alpha} and \eqn{\lambda}.
}
\details{
The expected value of the Z-PSS distribution only exists for \eqn{\alpha} values strictly greater than 2.
}
\examples{
zpssMean(2.5, 1.3)
zpssMean(2.5, 1.3, TRUE)
}
