% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpssMoments.R
\name{zpssMoments}
\alias{zpssMoments}
\title{Distribution Moments.}
\usage{
zpssMoments(k, alpha, lambda, isTruncated = FALSE, tolerance = 10^(-4))
}
\arguments{
\item{k}{Order of the moment to compute.}

\item{alpha}{Value of the \eqn{\alpha} parameter (\eqn{\alpha > k + 1}).}

\item{lambda}{Value of the \eqn{\lambda} parameter (\eqn{\lambda \geq 0}).}

\item{isTruncated}{Logical; if TRUE, the truncated version of the distribution is returned.}

\item{tolerance}{Tolerance used in the calculations (default = \eqn{10^{-4}}).}
}
\value{
A positive real value corresponding to the k-th moment of the distribution.
}
\description{
General function to compute the k-th moment of the Z-PSS distribution. Note that the k-th moment
exists if and only if  \eqn{\alpha > k + 1}.
}
\details{
The k-th moment of the Z-PSS distribution is finite for \eqn{\alpha} values strictly greater than \eqn{k + 1}.
It is computed calculating the partial sums of the serie, and it stops when two consecutive
partial sums differs less than the \code{tolerance} value. The last partial sum is returned.
}
\examples{
zpssMoments(3, 4.5, 1.3)
zpssMoments(3, 4.5, 1.3,  1*10^(-3))
}
