% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpeFit.R
\name{zpeFit}
\alias{zpeFit}
\title{ZPE parameters estimation.}
\usage{
zpeFit(data, init_alpha, init_beta, ...)
}
\arguments{
\item{data}{Matrix of count data.}

\item{init_alpha}{Initial value of \eqn{\alpha} parameter (\eqn{\alpha > 1}).}

\item{init_beta}{Initial value of \eqn{\beta} parameter (\eqn{\beta \in [ -\infity, +\infity]}).}

\item{...}{Further arguments to the generic functions. In case of the function \emph{moezipfR.fit}
the extra arguments are passing to the \link{optim} function.}
}
\value{
Returns a \emph{moezipfR} object composed by the maximum likelihood parameter estimations,
their standard deviation, their confidence intervals and the log-likelihood value.
}
\description{
For a given count data set,  usually of the type of ranking data or frequencies of frequencies data, estimates the parameters of the MOEZipf distribution by means of the maximum likelihood
method.
}
\details{
The argument \code{data} is a matrix where, for each row, the first column contains a count,
and the second column contains its corresponding frequency.

The log-likelihood function is computed by means of the following equation:


The function \emph{\link{optim}} is used to estimate the parameters.
}
\examples{
data <- rmoezipf(100, 2.5, 1.3)
data <- zipfExtR_getDataMatrix(data)
obj <- zpeFit(data, 1.001, 0.001)
}
\seealso{
\code{\link{zipfExtR_getDataMatrix}}, \code{\link{moezipf_getInitialValues}}.
}
