% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zpss.R
\name{zpss}
\alias{zpss}
\alias{dzpss}
\alias{dzpss}
\alias{pzpss}
\title{The Zipf-Poisson Stop Sum Distribution (Z-PSS).}
\usage{
dzpss(x, alpha, lambda, log = FALSE, isTruncated = FALSE)

pzpss(q, alpha, lambda, log.p = FALSE, lower.tail = TRUE,
  isTruncated = FALSE)
}
\arguments{
\item{x, q}{Vector of positive integer values.}

\item{alpha}{Value of the \eqn{\alpha} parameter (\eqn{\alpha > 1} ).}

\item{lambda}{Value of the \eqn{\lambda} parameter (\eqn{\lambda >= 0} ).}

\item{log, log.p}{Logical; if TRUE, probabilities p are given as log(p).}

\item{isTruncated}{Logical; if TRUE, the truncated version of the distribution is returned.}

\item{lower.tail}{Logical; if TRUE (default), probabilities are \eqn{P[X \leq x]}, otherwise, \eqn{P[X > x]}.}
}
\description{
Probability Mass Function, Cumulative Function of the Z-PSS distribution
with parameter \eqn{\alpha} and \eqn{\lambda}.
}
